admin_ui/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pyyaml==6.0.1
docker==7.0.0
pydantic==2.6.1
httpx==0.27.0
urllib3<2
requests==2.31.0
psutil==5.9.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
structlog==25.5.0
websockets==12.0
aiohttp==3.9.3
prometheus_client==0.19.0
